Added EntityOrData union
