+++ src/HotChocolate/Core/src/Abstractions/NameStringExtensions.cs
@@ -1,0 +1,22 @@
+using System;
+using HotChocolate.Properties;
+
+namespace HotChocolate
+{
+    public static class NameStringExtensions
+    {
+        public static NameString EnsureNotEmpty(
+            in this NameString name,
+            string argumentName)
+        {
+            if (name.IsEmpty)
+            {
+                throw new ArgumentException(
+                    AbstractionResources.Name_MustNotBeEmpty,
+                    argumentName);
+            }
+
+            return name;
+        }
+    }
+}
